[*] bring back tests + benchmarks + fuzzing + examples
- added convenience justfile rule to test like CI (w/ different feature combinations),
- removed executable_heap feature (heap is always executable now).
- fixed check_stack_guard functionality w/ stack cookie (currently broken in main).
- fixed logging to get proper output_data_region address (host vs. guest address space when
in process vs. in hypervisor).
- modified print_output to return the message length.

+++ src/hyperlight_guest/src/entrypoint.rs
@@ -86,7 +86,7 @@ pub extern "win64" fn entrypoint(peb_address: u64, seed: u64, max_log_level: u64
 
         // The guest sets the address to a "guest function dispatch" function, which is a function
         // that is called by the host to dispatch calls to guest functions.
-        (*PEB).set_guest_function_dispatch_ptr(dispatch_function as u64);
+        (*PEB).set_guest_function_dispatch_ptr(dispatch_function as usize as u64);
 
         // Set up the guest heap
         HEAP_ALLOCATOR
